[kitsune][test] Cleanup frontend tests
  - Remove unnecessary includes - especially to standard library headers. In
    places where printf was used, just use some declared external function.
    Similary, NULL has just been defined in the test files. It is not clear how
    much time this will save, but there is no point parsing more than needed.

  - Move the check lines above the corresonding code instead of below. This is
    consistent with all the other tests, and is more readable as well.

  - Where possible, some tests have been simplified. In many cases, explanations
    were added describing what the test is intended to check.

+++ kitsune/test/frontend/c/builtins.c
@@ -9,11 +9,9 @@
 // RUN: %kitcc -O0 -S -emit-llvm -o - %s %sysroot | FileCheck %s
 // RUN: %kitcc --tapir=nolo -O0 -S -emit-llvm -o - %s %sysroot | FileCheck %s
 
-#include <stddef.h>
-
 // CHECK-LABEL: allocate_c
 // CHECK: call ptr addrspace(67) @llvm.kit.mobile.alloc(i64 %{{.+}})
-void *__attribute__((kitsune_mobile)) allocate_c(size_t n) {
+void *__attribute__((kitsune_mobile)) allocate_c(unsigned long n) {
   return kitsune_mobile_alloc(n);
 }
